+++ dragpoint.cpp
@@ -596,7 +596,7 @@ STDMETHODIMP DragPoint::InterfaceSupportsErrorInfo(REFIID riid)
       &IID_IControlPoint,
    };
 
-   for (int i=0;i<sizeof(arr)/sizeof(arr[0]);i++)
+   for (size_t i=0;i<sizeof(arr)/sizeof(arr[0]);i++)
    {
       if (InlineIsEqualGUID(*arr[i],riid))
          return S_OK;
