+++ textbox.cpp
@@ -192,7 +192,7 @@ STDMETHODIMP Textbox::InterfaceSupportsErrorInfo(REFIID riid)
 		&IID_ITextbox,
 	};
 
-	for (int i=0;i<sizeof(arr)/sizeof(arr[0]);i++)
+	for (size_t i=0;i<sizeof(arr)/sizeof(arr[0]);i++)
 	{
 		if (InlineIsEqualGUID(*arr[i],riid))
 			return S_OK;
